Refactoring legacy functions into its own module
